+++ tests/ExpectationTest.php
@@ -4,7 +4,6 @@ namespace tests;
 
 use GuzzleHttp\Psr7\Response;
 use BlastCloud\Guzzler\{Expectation, UsesGuzzler};
-use PHPUnit\Framework\AssertionFailedError;
 use PHPUnit\Framework\TestCase;
 
 class ExpectationTest extends TestCase
@@ -74,57 +73,6 @@ class ExpectationTest extends TestCase
 
         $this->guzzler->expects($this->never())
             ->something('/a-url');
-    }
-
-    public function testEachConvenienceVerbMethodDoesntErr()
-    {
-        $expectation = $this->guzzler->expects($this->never());
-
-        foreach (Expectation::VERBS as $verb) {
-            $expectation->$verb('/a-url');
-        }
-    }
-
-    public function testSynchronous()
-    {
-        $this->guzzler->queueMany(new Response(), 2);
-
-        $this->guzzler->expects($this->once())
-            ->synchronous();
-
-        // First test a passing call
-        $this->client->get('/a-url');
-
-        // Now test a failing call
-        $this->client->getAsync('/anywhere')->wait();
-
-        $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessageRegExp("/\bsynchronous\b/");
-
-        $this->guzzler->assertLast(function (Expectation $e) {
-            return $e->synchronous();
-        });
-    }
-
-    public function testAsynchronous()
-    {
-        $this->guzzler->queueMany(new Response(), 2);
-
-        $this->guzzler->expects($this->once())
-            ->asynchronous();
-
-        // First test a passing call
-        $this->client->getAsync('/woeiwj')->wait();
-
-        // Now test a failing call
-        $this->client->get('/somewhere');
-
-        $this->expectException(AssertionFailedError::class);
-        $this->expectExceptionMessageRegExp("/\bsynchronous\b/");
-
-        $this->guzzler->assertLast(function (Expectation $e) {
-            return $e->asynchronous();
-        });
     }
 
     public function testFailureWhenWithNotFound()