+++ tests/testFiles/macros.php
@@ -1,0 +1,7 @@
+<?php
+
+use BlastCloud\Guzzler\Expectation;
+
+Expectation::macro('fromFile', function (Expectation $e, $url) {
+    return $e->post($url);
+});